changed: put movieset videodatabase methods in separate file
